move util functions to separat file for testing
